extract Phi data model to new file
